Evolution files and JDBC Access (#4)
* Setup PostgreSQL database and JDBC

Restructured the Project to include the Play Framework and added an endpoint that returns a status JSON

Database has been set up and with it, getInventory() was created alongside a transactor using Doobie and Cats to work as JDBC with PostgresSQL.

Made PostgreSQL part of the project without needing pgAdmin. Created SQL evolution and accessing for database. README is still under work.

Deleting .g8/ folder

* Added ProductDAO service and updated the README.

* Fixes and refinement for the discussed items.

* More fixes that were previously discussed