CRUD Addition and completed endpoints (#5)
* Setup PostgreSQL database and JDBC

Restructured the Project to include the Play Framework and added an endpoint that returns a status JSON

Database has been set up and with it, getInventory() was created alongside a transactor using Doobie and Cats to work as JDBC with PostgresSQL.

Made PostgreSQL part of the project without needing pgAdmin. Created SQL evolution and accessing for database. README is still under work.

Deleting .g8/ folder

* Added ProductDAO service and updated the README.

* Defined the CRUD and completed making all the endpoints. Refined README, and fixed small issues.

* Code quality and improvements for what was discussed.

* More Improvements for code.

* Added all the tests.

* Updated README file to include a note below sbt run command on how to start the database server.

* Improved tests as discussed

* Changed test names to make it easier to understand